Downloads/notes/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
cloudinary==1.36.0
httpx==0.26.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
python-socketio==5.10.0
websockets==12.0
aiofiles==23.2.1
cryptography==42.0.2
# Firebase Admin SDK for Firestore
firebase-admin==6.5.0
# Optional: pytesseract (for OCR - requires Tesseract OCR to be installed on system)
# pytesseract
# Docker SDK for container management
docker==7.0.0
# For WebRTC signaling (optional - can use external service)
# aiortc
